back/av_db/requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
django-cors-headers==4.4.0
django-extensions==3.2.3
django-redis==5.4.0
djangorestframework==3.15.2
idna==3.8
mysqlclient==2.2.4
pandas==2.2.3
python-dotenv==1.0.1
redis==5.0.8
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
boto3==1.36.22
mypy==1.15.0
